% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.axis_scaler}
\alias{.axis_scaler}
\title{Inverse of .variable_scaler for secondary-axis labeling}
\usage{
.axis_scaler(y1, yaxis1, yaxis2)
}
\arguments{
\item{y1}{Values on the primary scale.}

\item{yaxis1}{Range (min, max) of the primary axis.}

\item{yaxis2}{Range (min, max) of the secondary axis.}
}
\value{
\code{y1} linearly mapped into secondary-axis coordinates.
}
\description{
Inverse of .variable_scaler for secondary-axis labeling
}
\keyword{internal}
