% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.variable_scaler}
\alias{.variable_scaler}
\title{Rescale a secondary-axis variable into primary-axis coordinates}
\usage{
.variable_scaler(y2, yaxis1, yaxis2)
}
\arguments{
\item{y2}{Values on the secondary scale.}

\item{yaxis1}{Range (min, max) of the primary axis.}

\item{yaxis2}{Range (min, max) of the secondary axis.}
}
\value{
\code{y2} linearly mapped into primary-axis coordinates.
}
\description{
Used by the distribution plots (LCD/LRD/TRP) to overlay a frequency
polygon on a bar chart with a different y-scale.
}
\keyword{internal}
