% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{ICC_plot}
\alias{ICC_plot}
\title{ICC function defin}
\usage{
ICC_plot(data, xvariable = c(-4, 4))
}
\arguments{
\item{data}{Exametrika output results}

\item{xvariable}{Specify the vector to set the drawing range at both ends of the x-axis}
}
\description{
This function, when provided with parameters, returns a logistic function
}
