% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04IRT.R
\name{LogisticModel}
\alias{LogisticModel}
\title{Four-Parameter Logistic Model}
\usage{
LogisticModel(x, a = 1, b, c = 0, d = 1)
}
\arguments{
\item{x}{ability parameter}

\item{a}{slope parameter}

\item{b}{locaiton parameter}

\item{c}{lower asymptote parameter}

\item{d}{upper asymptote parameter}
}
\description{
The four-parameter logistic model is a model where one additional
parameter d, called the upper asymptote parameter, is added to the
3PLM.
}
