% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICCtoTIC.R
\name{plotIIC_gg}
\alias{plotIIC_gg}
\title{Plot Item Characteristic Curves (IIC) from Exametrika}
\usage{
plotIIC_gg(data, xvariable = c(-4, 4))
}
\arguments{
\item{data}{Exametrika output results}

\item{xvariable}{Specify the vector to set the drawing range at both ends of the x-axis}
}
\description{
This function takes Exametrika output as input
and generates Item Characteristic Curves (IIC) using ggplot2.
The applicable analytical method is Item Response Theory (IRT).
}
