% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotIRP_gg}
\alias{plotIRP_gg}
\title{Plot Item Reference Profile (IRP) from Exametrika}
\usage{
plotIRP_gg(data)
}
\arguments{
\item{data}{Exametrika output results}
}
\description{
This function takes Exametrika output as input
and generates Item Reference Profile (IRP) using ggplot2.
The applicable analytical methods are Latent Class Analysis (LCA), Latent Rank Analysis (LRA), and Local Dependent Latent Rank Analysis (LDLRA).
}
