% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotLCD_gg}
\alias{plotLCD_gg}
\title{Plot Latnt Class Distribution (LCD) from Exametrika}
\usage{
plotLCD_gg(data, Num_Students = TRUE, title = TRUE)
}
\arguments{
\item{data}{Exametrika output results}

\item{Num_Students}{Display the number of students on the bar graph}

\item{title}{Toggle the presence of the title using TRUE/FALSE}
}
\description{
This function takes Exametrika output as input
and generates Latnt Class Distribution (LCD) using ggplot2.
The applicable analytical methods are Latent Class Analysis (LCA) and Bicluster Network Model (BINET).
}
