% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotLRD_gg}
\alias{plotLRD_gg}
\title{Plot Latnt Rank Distribution (LRD) from Exametrika}
\usage{
plotLRD_gg(data, Num_Students = TRUE, title = TRUE)
}
\arguments{
\item{data}{Exametrika output results}

\item{Num_Students}{Display the number of students on the bar graph}

\item{title}{Toggle the presence of the title using TRUE/FALSE}
}
\description{
This function takes Exametrika output as input
and generates Latnt Rank Distribution (LRD) using ggplot2.
The applicable analytical methods are Latent Rank Analysis (LRA), Biclustering,
Local Dependent Latent Rank Analysis (LDLRA), and Local Dependence Biclustering (LDB).
}
