% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICCtoTIC.R
\name{plotTIC_gg}
\alias{plotTIC_gg}
\title{Plot Test Information Curves (TIC) from Exametrika}
\usage{
plotTIC_gg(data, xvariable = c(-4, 4))
}
\arguments{
\item{data}{Exametrika output results}

\item{xvariable}{Specify the vector to set the drawing range at both ends of the x-axis}
}
\description{
This function takes Exametrika output as input
and generates Test Information Curves (TIC) using ggplot2.
The applicable analytical method is Item Response Theory (IRT).
}
