% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{combinePlots_gg}
\alias{combinePlots_gg}
\title{Display Multiple Plots Simultaneously from ggExametrika Output"}
\usage{
combinePlots_gg(plots, selectPlots = c(1:6))
}
\arguments{
\item{plots}{ggExametrika output plots}

\item{selectPlots}{In the vector, specify the range or indices you want to display.}
}
\description{
This function consolidates multiple plots generated from the ggExametrika function into a single display.
}
