% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04IRT.R
\name{ItemInformationFunc}
\alias{ItemInformationFunc}
\title{IIF for 4PLM}
\usage{
ItemInformationFunc(x, a = 1, b, c = 0, d = 1)
}
\arguments{
\item{x}{ability parameter}

\item{a}{slope parameter}

\item{b}{locaiton parameter}

\item{c}{lower asymptote parameter}

\item{d}{upper asymptote parameter}
}
\description{
Item Information Function for 4PLM
}
