% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arraytoLDPSR.R
\name{plotArray_gg}
\alias{plotArray_gg}
\title{Plot Array from Exametrika}
\usage{
plotArray_gg(
  data,
  Original = TRUE,
  Clusterd = TRUE,
  Clusterd_lines = TRUE,
  title = TRUE
)
}
\arguments{
\item{data}{Exametrika output results}

\item{Original}{plot original data}

\item{Clusterd}{plot Clusterd data}

\item{Clusterd_lines}{plot the red lines representing ranks and fields}

\item{title}{Presence or absence of a title.}
}
\description{
This function takes Exametrika output as input
and generates Rank Membership Profile (RMP) using ggplot2.
The applicable analytical methods are Latent Rank Analysis (LRA), Biclustering,
Local Dependent Latent Rank Analysis (LDLRA), and Local Dependence Biclustering (LDB).
}
