% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotCMP_gg}
\alias{plotCMP_gg}
\title{Plot Class Membership Profile (CMP) from Exametrika}
\usage{
plotCMP_gg(data)
}
\arguments{
\item{data}{Exametrika output results}
}
\description{
This function takes Exametrika output as input
and generates Class Membership Profile (CMP) using ggplot2.
The applicable analytical methods are Latent Class Analysis (LCA) and Bicluster Network Model (BINET).
}
