% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arraytoLDPSR.R
\name{plotFieldPIRP_gg}
\alias{plotFieldPIRP_gg}
\title{Plot FieldPIRP from Exametrika}
\usage{
plotFieldPIRP_gg(data)
}
\arguments{
\item{data}{Exametrika output results}
}
\description{
This function takes Exametrika output as input
and generates FieldPIRP using ggplot2.
The applicable analytical methods is Local Dependence Biclustering (LDB).
The warning message regarding NA values is displayed, but the behavior is normal.
}
