% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotFRP_gg}
\alias{plotFRP_gg}
\title{Plot Item Reference Profile (FRP) from Exametrika}
\usage{
plotFRP_gg(data)
}
\arguments{
\item{data}{Exametrika output results}
}
\description{
This function takes Exametrika output as input
and generates Field Reference Profile (FRP) using ggplot2.
The applicable analytical methods are Biclustering, Infinite Relational Model (IRM),
Local Dependence Biclustering (LDB), and Bicluster Network Model (BINET).
}
