% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotIRP_gg}
\alias{plotIRP_gg}
\title{Plot Item Reference Profile (IRP) from Exametrika}
\usage{
plotIRP_gg(data)
}
\arguments{
\item{data}{Exametrika output results}

\item{xvariable}{Specify the vector to set the drawing range at both ends of the x-axis}
}
\description{
This function takes Exametrika output as input
and generates Test Information Curves (IRP) using ggplot2.
The applicable analytical methods are Latent Class Analysis (LCA), Latent Rank Analysis (LRA),
Biclustering, and Local Dependent Latent Rank Analysis (LDLRA).
}
