% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotRMP_gg}
\alias{plotRMP_gg}
\title{Plot Rank Membership Profile (RMP) from Exametrika}
\usage{
plotRMP_gg(data)
}
\arguments{
\item{data}{Exametrika output results}
}
\description{
This function takes Exametrika output as input
and generates Rank Membership Profile (RMP) using ggplot2.
The applicable analytical methods are Latent Rank Analysis (LRA), Biclustering,
Local Dependent Latent Rank Analysis (LDLRA), and Local Dependence Biclustering (LDB).
}
