% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPtoCMPRMP.R
\name{plotTRP_gg}
\alias{plotTRP_gg}
\title{Plot Test Reference Profile (TRP) from Exametrika}
\usage{
plotTRP_gg(data, Num_Students = TRUE, title = T)
}
\arguments{
\item{data}{Exametrika output results}

\item{title}{Toggle the presence of the title using TRUE/FALSE}
}
\description{
This function takes Exametrika output as input
and generates Test Reference Profile (TRP) using ggplot2.
The applicable analytical methods are Latent Class Analysis (LCA), Latent Rank Analysis (LRA), Biclustering,
Infinite Relational Model (IRM), Local Dependence Biclustering (LDB), and Bicluster Network Model (BINET).
}
